// src/app/enums.ts
export enum Categories {
  Grammatik = "Grammatik",
  Rechtschreibung = "Rechtschreibung",
  Textsorten = "Textsorten",
}

export enum Topics {
  Wortarten = "Wortarten",
  Satzarten = "Satzarten & Satzglieder",
  Zeiten = "Zeiten richtig anwenden",
  Faelle = "Fälle & Präpositionen",
  Beistrich = "Beistrichsetzung",
  Quiz = "Quiz",
}

export enum GrammarCase {
  Akkusativ = "Akkusativ",
  Dativ = "Dativ",
}